+++ config/routes.rb
@@ -137,6 +137,7 @@ Rails.application.routes.draw do
   get 'research-topics/accepted/didgeridoo-a-potentially-novel-intervention-for-sleep-apnea', to: 'research_topics#sleep_apnea_didgeridoo', as: 'sleep_apnea_didgeridoo'
   get 'research-topics/accepted/unilateral-hypoglossal-nerve-stimulation-sleep-apnea-treatment', to: 'research_topics#sleep_apnea_hypoglossal_stimulation', as: 'sleep_apnea_hypoglossal_stimulation'
   get 'research-topics/accepted/women-with-sleep-apnea-increased-risk-for-heart-disease-with-age', to: 'research_topics#sleep_apnea_women_heart_disease', as: 'sleep_apnea_women_heart_disease'
+  get 'research-topics/accepted/atrial-fibrillation-and-its-links-to-sleep-apnea', to: 'research_topics#sleep_apnea_afib', as: 'sleep_apnea_afib'
 
   # Surveys
   resources :surveys do